first pass on code reorg
